ファイルわける
