+++ src/ts/issue.ts
@@ -1,0 +1,8 @@
+export class Issue {
+    constructor(
+        public repo: string,
+        public number: string,
+        public state: string,
+        public assignee: { avatar_url: string }
+    ) {}
+}
